feat: 가게별 평점 소숫점 수정 (#184)
* feat: addExposedHeader 속성 추가

* fix: domain 주소 수정

* fix: domain 주소 수정

* fix: domain 주소 수정

* feat: java doc 추가, Store 엔티티, DTO추가

* rename: 프론트엔드와의 JSON 필드명 맞춤

* fix : localhost:3000 Origin 추가

* fix : https://deploymentprac38.tk alllow-origin 추가

* feat : Menu : Store N:1 관계 추가

* fix: 비밀번호 검사 x

* fix: domain 속성 제거

* fix: Access-Control-Allow-Header 속성 수정

* fix: Access-Control-Allow-Header 원상 복구

* fix: ExposedHeader 추가

* fix: cors()설정 enable() (안되면 롤백)

* fix: cors()설정 롤백

* fix : 헤더 속성 변경

* fix : 헤더 속성 변경

* fix: ExposedHeader 속성 변경

* feat: 비밀번호 찾기 이메일 인증코드 추가

* 접근권한 설정

* fix: .gitignore에 server, local 프로파일 추가

* fix: build.gradle 외부라이브러리 수정

* fix: Local -> Store 참조 시 null 고려

* fix: 로컬 Logout 토큰 삭제 방법을 security -> mvc로 변경

* fix: 로컬 로그아웃 URL 수정

* fix: 로컬 로그아웃 로직, 토큰 메서드 수정

* fix: refresh_token unique 삭제, phone unique 추가

* fix: 예외 처리 수정

* fix: 예외 처리 수정

* feat: 필터에서의 예외 처리 추가

* fix: 필터 예외 처리 한글 인코딩(UTF-8) 지정

* feat: S3 이미지 저장 기능 추가

* feat: S3에 이미지 저장 기능 추가

* fix: S3에 저장되는 방식 변경

* fix: S3에 Base64 디코딩 해서 저장

* fix: S3에 Base64 수정

* feat : store, review S3에 이미지 CUD 기능 추가

* feat: store, review S3에 이미지 CUD 기능 추가

* fix: 이미지 삭제 경로 수정

* feat: menu 이미지 CUD 기능 추가

* feat: 이미지 포맷 예외처리 추가, kakao 마이페이지 리턴에 주문내역들 추가

* feat: 가게별 평점 기능 추가

* feat: 가게별 평점 소숫점 수정

+++ src/main/java/com/main_39/Spring/store/dto/StoreResponseDto.java
@@ -22,7 +22,7 @@ public class StoreResponseDto {
         private String storeTag;
 
         private int totalReview;
-        private int totalGrade;
+        private double totalGrade;
         private int totalComment;
 
 //        private List<ReviewResponseDto> reviews;
